Generalize abstract `InterfaceCommand`
